+++ CHANGELOG.md
@@ -7,6 +7,10 @@
   types, as well as the (aggregate) amount of deductions related to
   foreign income.
 
+- Implement the Foreign Income Tax Offset Limit.  The limit will be
+  calculated and the `foreignTaxOffset` field in the `TaxReturnInfo`
+  will be clamped to it.
+
 ## Older versions
 
 See Git commit history